* MVVC Models now use Afterthought for notification raising.  * File copier now skips files that already exist at destination.  * Improved AboutWindow appearance.    * Working links to GPL and Project website.    * Trademark notifications about "BlackBerry".  * This revision of MusicSync.Desktop has a bug which causes app crash after copying (but not deleting) a single file.

+++ MusicSync.Desktop/Threading/FileCopier.cs
@@ -49,6 +49,15 @@ namespace WifiSyncDesktop.Threading
         public override void Work(FileOperation job)
         {
             Directory.CreateDirectory(Path.GetDirectoryName(job.Destination));
+            if(File.Exists(job.Destination))
+            {
+                // Check if the sizes are the same and skip copying if it is.
+                FileInfo srcInfo = new FileInfo(job.Source);
+                FileInfo destInfo = new FileInfo(job.Destination);
+
+                if (srcInfo.Length == destInfo.Length) return;
+            }
+
             File.Copy(job.Source, job.Destination, true);
         }
